make packet id 2 bytes instead of 4

+++ src/main/java/com/github/mkorman9/game/dto/packet/Response.java
@@ -1,5 +1,5 @@
 package com.github.mkorman9.game.dto.packet;
 
 public interface Response {
-    int packetId();
+    short packetId();
 }
